Create Read and Rewrite Binary File Method Implement Code in Review Application - feature: Approve Marriage,reject Marriage, verify Marriage, Search BinaryFile

+++ src/main/java/com/example/kazi_office_and_marriage_register_office/yasin/ReviewApplicationController.java
@@ -80,7 +80,7 @@ public class ReviewApplicationController
             File f = new File("SavedApplications.bin");
             FileOutputStream fos = null;
             ObjectOutputStream oos = null;
-            if(f.exists()){
+            if(f.exists() && f.length()>0){
                 fos = new FileOutputStream(f,true);
                 oos = new appendableObjectOutputStream(fos);
             }
